Error when updating ticks when comp class is not in progress

+++ backend/src/main/kotlin/se/scoreboard/api/TickController.kt
@@ -2,19 +2,24 @@ package se.scoreboard.api
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.domain.Pageable
+import org.springframework.http.HttpStatus
 import org.springframework.messaging.simp.SimpMessagingTemplate
 import org.springframework.web.bind.annotation.*
 import se.scoreboard.data.domain.Contender
 import se.scoreboard.data.domain.extension.getQualificationScore
 import se.scoreboard.data.domain.extension.getTotalScore
+import se.scoreboard.data.domain.extension.isInProgress
 import se.scoreboard.dto.ScoreboardListItemDto
 import se.scoreboard.dto.ScoreboardPushItemDto
 import se.scoreboard.dto.TickDto
+import se.scoreboard.exception.WebException
 import se.scoreboard.service.ContenderService
 import se.scoreboard.service.TickService
+import javax.transaction.Transactional
 
 @RestController
 @CrossOrigin
+@Transactional
 @RequestMapping("/api")
 class TickController @Autowired constructor(
         val tickService: TickService,
@@ -41,8 +46,13 @@ class TickController @Autowired constructor(
 
     @PostMapping("/tick")
     fun createTick(@RequestBody tick : TickDto): TickDto {
+        val contender = contenderService.fetchEntity(tick.contenderId!!)
+        if(!contender.compClass!!.isInProgress()) {
+            throw WebException(HttpStatus.FORBIDDEN, "The competition is not in progress");
+        }
+
         val newTick = tickService.create(tick)
-        broadcast(contenderService.fetchEntity(tick.contenderId!!))
+        broadcast(contender)
         return newTick
     }
 
@@ -50,14 +60,23 @@ class TickController @Autowired constructor(
     fun updateTick(
             @PathVariable("id") id: Int,
             @RequestBody tick : TickDto): TickDto {
+        val contender = contenderService.fetchEntity(tick.contenderId!!)
+        if(!contender.compClass!!.isInProgress()) {
+            throw WebException(HttpStatus.FORBIDDEN, "The competition is not in progress");
+        }
         val newTick =tickService.update(id, tick)
-        broadcast(contenderService.fetchEntity(tick.contenderId!!))
+        broadcast(contender)
         return newTick
     }
 
     @DeleteMapping("/tick/{id}")
     fun deleteTick(@PathVariable("id") id: Int) {
-        val tick = tickService.delete(id)
-        broadcast(contenderService.fetchEntity(tick.contenderId!!))
+        var tick = tickService.fetchEntity(id)
+        val contender = tick.contender
+        if(!contender!!.compClass!!.isInProgress()) {
+            throw WebException(HttpStatus.FORBIDDEN, "The competition is not in progress");
+        }
+        tickService.delete(id)
+        broadcast(contender!!)
     }
 }
